ModalParent sty-com refactor
