+++ src/components/modal/ModalStyling.js
@@ -4,6 +4,34 @@ import zoomOut from 'react-animations/lib/zoom-out';
 import { MdClose } from 'react-icons/md';
 
 
+export const Container = styled.div`
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 10vh;
+    
+    `;
+    
+export const Button = styled.button`
+    min-width: 150px;
+    padding: 7px 14px;
+    border-radius: 20px;
+    border: 2px solid blue;
+    backgroud: transparent;
+    color: blue;
+    font-size: 16px;
+    cursor: pointer;
+
+    &:hover { 
+        background-color: blue;
+        color: white;
+        transform: scale(1.2)
+    }
+
+`;
+
+
+
 
 export const Background = styled.div`
     width: 100%;
